Replace ResourceChartingPage to Charting component
Added Extra to PageTabsProps and Tab
