+++ src/components/Charting/ChartingList/styles.ts
