+++ src/uberComponents/ResourceDetailPage/types.ts
@@ -4,11 +4,11 @@ import { SearchParams, WithId } from '@beda.software/fhir-react';
 
 import { RecordType } from '../ResourceListPage/types';
 
-export interface Tab<R extends Resource> {
+export type Tab<R extends Resource, Extra = unknown> = {
     label: string;
     path?: string;
     component: (context: RecordType<R>) => JSX.Element;
-}
+} & Extra;
 
 export interface DetailPageProps<R extends Resource> {
     resourceType: R['resourceType'];
@@ -18,6 +18,6 @@ export interface DetailPageProps<R extends Resource> {
     extractPrimaryResource?: (bundle: Bundle<R>) => WithId<R>;
 }
 
-export interface PageTabsProps<R extends Resource> {
-    tabs: Array<Tab<WithId<R>>>;
-}
+export type PageTabsProps<R extends Resource, Extra = unknown> = {
+    tabs: Array<Tab<WithId<R>, Extra>>;
+};
