+++ src/uberComponents/index.ts
@@ -1,4 +1,1 @@
 export * from './ResourceListPage';
-export * from './ResourceChartingPage';
-export * from './ResourceChartingPage/types';
-export * from './ResourceChartingPage/ChartingList';
